+++ mod/common/src/main/kotlin/dev/wanderia/netlib/payload/api/PayloadChannel.kt
@@ -1,0 +1,14 @@
+/*
+ * Copyright (C) 2024 Wanderia - All Rights Reserved
+ *
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at https://mozilla.org/MPL/2.0/.
+ */
+package dev.wanderia.netlib.payload.api
+
+@Deprecated(
+    replaceWith = ReplaceWith("dev.wanderia.netlib.payload.PayloadChannel"),
+    message = "Moved Package",
+)
+public typealias PayloadChannel = dev.wanderia.netlib.payload.PayloadChannel