+++ mod/common/src/main/kotlin/dev/wanderia/netlib/payload/api/SerializedPayloadConfiguration.kt
@@ -1,0 +1,18 @@
+/*
+ * Copyright (C) 2024 Wanderia - All Rights Reserved
+ *
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at https://mozilla.org/MPL/2.0/.
+ */
+package dev.wanderia.netlib.payload.api
+
+import kotlinx.serialization.ExperimentalSerializationApi
+
+@ExperimentalSerializationApi
+@Deprecated(
+    replaceWith = ReplaceWith("dev.wanderia.netlib.payload.SerializedPayloadConfiguration"),
+    message = "Moved Package",
+)
+public typealias SerializedPayloadConfiguration<T> =
+    dev.wanderia.netlib.payload.SerializedPayloadConfiguration<T>